Username/password fields are now named correctly
This was originally done to try and stop Chrome autofilling them with the wrong data.

However, Chrome ignores all the autocomplete standards and does whatever it wants anyway—even with this workaround in place.

So, since this is unlikely to be fixed (see https://bugs.chromium.org/p/chromium/issues/detail?id=914451 which has been open for three YEARS), we may as well at least have readable code!

Closes #128.